pyproject.toml
[project]
name = "api-challenge"
version = "0.1.0"
description = "My response to Ali's welcome Api challenge"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-environ>=0.11.2",
    "djangorestframework>=3.15.2",
    "drf-nested-routers>=0.94.1",
    "pillow>=11.0.0",
    "psycopg>=3.2.3",
]
